+++ tests/integration/cases/tuple_pair_record_field/Zig_heterogeneous_strategy_record@v0_12.zig
@@ -1,0 +1,11 @@
+const Record0 = struct { call: []const u8, args: struct { i64, []const u8 } };
+pub fn main() void {
+    const my_data = Record0{
+        .call = "send",
+        .args = .{
+            1,
+            "email",
+        },
+    };
+    _ = my_data;
+}